requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.9.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-WTF==1.2.1
google==3.0.0
idna==3.7
imbalanced-learn==0.12.3
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lightgbm==4.3.0
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.9.4
pytz==2024.1
regex==2024.5.15
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
soupsieve==2.5
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.1.2
